pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evalscope"
version = "0.1.0"
description = "Local-first eval workflows for LLM apps"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.30.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
evalscope = "evalscope.cli:main"

[tool.setuptools]
packages = ["evalscope"]
